+++ src/quits/__init__.py
@@ -1,0 +1,4 @@
+"""QUITS package."""
+
+from .api import *  # noqa: F403
+
