+++ src/quits/decoder/__init__.py
@@ -1,0 +1,24 @@
+"""Decoder utilities and sliding window implementations."""
+
+from .base import (
+    detector_error_model_to_matrix,
+    dict_to_csc_matrix_column_row,
+    dict_to_csc_matrix_row_column,
+    spacetime,
+)
+from .bplsd import sliding_window_bplsd_circuit_mem, sliding_window_bplsd_phenom_mem
+from .bposd import sliding_window_bposd_circuit_mem, sliding_window_bposd_phenom_mem
+from .sliding_window import sliding_window_circuit_mem, sliding_window_phenom_mem
+
+__all__ = [
+    "detector_error_model_to_matrix",
+    "dict_to_csc_matrix_column_row",
+    "dict_to_csc_matrix_row_column",
+    "spacetime",
+    "sliding_window_phenom_mem",
+    "sliding_window_bposd_phenom_mem",
+    "sliding_window_bplsd_phenom_mem",
+    "sliding_window_circuit_mem",
+    "sliding_window_bposd_circuit_mem",
+    "sliding_window_bplsd_circuit_mem",
+]
